Validate Initial Payload, Design Tweak and Fixed the Left Menu State
The new obj.length verifies if there is anything in the payload that comes from Holochain once the page is loaded. This is important because initially there is nothing in the draft.

I have also removed some classes such as the flex ones and now the date and title goes into a separate line.

Ultimately the RefreshDraft function was updated in order to fix the issue with the event. After understanding (actually remembering) how the DOM and events worked, I ended up binding the event into the GetDraft in order to access to the correct property that fired the event.